requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.54.3
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.4
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
schedule==1.2.0
Werkzeug==3.0.1